Fixed some issues with csv_reader sending longs to DataTests

+++ netCDF_Instrument/csv_readers.py
@@ -117,7 +117,7 @@ class House(NetCDFWriter):
         self.vstore.latitude = self.latitude
         self.vstore.longitude = self.longitude
         #Tests#
-        self.vstore.pressure_qc_data = DataTests.run_tests(self.pressure_data,0)
+        self.vstore.pressure_qc_data = DataTests.run_tests(self.pressure_data.astype(np.double), 0)
         self.write_netCDF(self.vstore, len(self.pressure_data))
 
 
@@ -214,7 +214,7 @@ class Leveltroll(NetCDFWriter):
         self.vstore.latitude = self.latitude
         self.vstore.longitude = self.longitude
         #Tests#
-        self.vstore.pressure_qc_data = DataTests.run_tests(self.pressure_data,0)
+        self.vstore.pressure_qc_data = DataTests.run_tests(self.pressure_data.astype(np.double),0)
         self.write_netCDF(self.vstore, len(self.pressure_data))
